Moving localize methods out of utils
