+++ module/essence20.mjs
@@ -10,7 +10,8 @@ import { Essence20ItemSheet } from "./sheets/item-sheet.mjs";
 import { highlightCriticalSuccessFailure } from "./chat.mjs";
 import { E20 } from "./helpers/config.mjs";
 import { preloadHandlebarsTemplates } from "./helpers/templates.mjs";
-import { getNumActions, performPreLocalization } from "./helpers/utils.mjs";
+import { getNumActions } from "./helpers/utils.mjs";
+import { performPreLocalization } from "./helpers/localize.mjs";
 import { migrateWorld } from "./migration.mjs";
 
 function registerSystemSettings() {
